Move Warning logic into Post Comments Form component
