+++ packages/block-library/src/post-comments-form/edit.js
@@ -9,14 +9,8 @@ import classnames from 'classnames';
 import {
 	AlignmentControl,
 	BlockControls,
-	Warning,
 	useBlockProps,
-	store as blockEditorStore,
 } from '@wordpress/block-editor';
-import { Button } from '@wordpress/components';
-import { useEntityProp, store as coreStore } from '@wordpress/core-data';
-import { __, _x, sprintf } from '@wordpress/i18n';
-import { useSelect } from '@wordpress/data';
 
 /**
  * Internal dependencies
@@ -30,71 +24,12 @@ export default function PostCommentsFormEdit( {
 } ) {
 	const { textAlign } = attributes;
 	const { postId, postType } = context;
-	const [ commentStatus, setCommentStatus ] = useEntityProp(
-		'postType',
-		postType,
-		'comment_status',
-		postId
-	);
+
 	const blockProps = useBlockProps( {
 		className: classnames( {
 			[ `has-text-align-${ textAlign }` ]: textAlign,
 		} ),
 	} );
-
-	const isSiteEditor = postType === undefined || postId === undefined;
-
-	const { defaultCommentStatus } = useSelect(
-		( select ) =>
-			select( blockEditorStore ).getSettings()
-				.__experimentalDiscussionSettings
-	);
-
-	const postTypeSupportsComments = useSelect( ( select ) =>
-		postType
-			? !! select( coreStore ).getPostType( postType )?.supports.comments
-			: false
-	);
-
-	let warning = false;
-	let actions;
-	let showPlaceholder = true;
-
-	if ( ! isSiteEditor && 'open' !== commentStatus ) {
-		if ( 'closed' === commentStatus ) {
-			warning = __(
-				'Post Comments Form block: Comments are not enabled for this item.'
-			);
-
-			actions = [
-				<Button
-					key="enableComments"
-					onClick={ () => setCommentStatus( 'open' ) }
-					variant="primary"
-				>
-					{ _x(
-						'Enable comments',
-						'action that affects the current post'
-					) }
-				</Button>,
-			];
-			showPlaceholder = false;
-		} else if ( ! postTypeSupportsComments ) {
-			warning = sprintf(
-				/* translators: 1: Post type (i.e. "post", "page") */
-				__(
-					'Post Comments Form block: Comments are not enabled for this post type (%s).'
-				),
-				postType
-			);
-			showPlaceholder = false;
-		} else if ( 'open' !== defaultCommentStatus ) {
-			warning = __(
-				'Post Comments Form block: Comments are not enabled.'
-			);
-			showPlaceholder = false;
-		}
-	}
 
 	return (
 		<>
@@ -107,11 +42,7 @@ export default function PostCommentsFormEdit( {
 				/>
 			</BlockControls>
 			<div { ...blockProps }>
-				{ warning && (
-					<Warning actions={ actions }>{ warning }</Warning>
-				) }
-
-				{ showPlaceholder ? <CommentsForm /> : null }
+				<CommentsForm postId={ postId } postType={ postType } />
 			</div>
 		</>
 	);
